Fix symbols with nets disappearing on move
Symbols that had nets in them that interacted with other nets would
disappear while moving. This was happening because the netList
generation did not explicitly check to make sure that the net wasn't a
top-level net. A consequence of this change will be that nets within
symbols will not merge/split properly. In most cases, this should be OK.
The user can ungroup the symbol if they desire that functionality of
those nets.